requirements.txt
python-dotenv==1.0.0
astrapy==0.7.4
pandas>=2.0.3
openai>=1.0.0
python-jose==3.3.0
requests==2.31.0
streamlit>=1.31.1
numpy>=2.1.0
plotly>=5.0.0